+++ modules/datadict/jsx/dataDictIndex.js
@@ -1,3 +1,4 @@
+import {createRoot} from 'react-dom/client';
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import Loader from 'Loader';
@@ -212,15 +213,16 @@ class DataDictIndex extends Component {
 }
 
 DataDictIndex.propTypes = {
-    dataURL: PropTypes.string.isRequired,
+  dataURL: PropTypes.string.isRequired,
+  fieldsURL: PropTypes.string,
 };
 
 window.addEventListener('load', () => {
-  ReactDOM.render(
-      <DataDictIndex
-        dataURL={`${loris.BaseURL}/datadict/?format=binary`}
-        fieldsURL={`${loris.BaseURL}/datadict/fields`}
-      />,
-      document.getElementById('lorisworkspace')
+  const root = createRoot(document.getElementById('lorisworkspace'));
+  root.render(
+    <DataDictIndex
+      dataURL={`${loris.BaseURL}/datadict/?format=binary`}
+      fieldsURL={`${loris.BaseURL}/datadict/fields`}
+    />
   );
 });