+++ modules/electrophysiology_browser/jsx/components/electrophysiology_session_panels.js
@@ -101,6 +101,7 @@ FilePanel.propTypes = {
     id: PropTypes.string,
     title: PropTypes.string,
     data: PropTypes.array,
+    children: PropTypes.array,
 };
 
 FilePanel.defaultProps = {