+++ modules/mri_violations/jsx/protocolModal.js
@@ -2,6 +2,7 @@ import Modal from 'Modal';
 import StaticDataTable from 'StaticDataTable';
 
 import React, {useEffect, useState} from 'react';
+import PropTypes from 'prop-types';
 
 
 /**
@@ -20,6 +21,9 @@ function ProtocolModal(props) {
         default: return null;
     }
 }
+ProtocolModal.propTypes = {
+    Type: PropTypes.string,
+};
 
 /**
  * Return a modal window to display a MRI violations protocol
@@ -247,6 +251,11 @@ function ProtocolViolationModal(props) {
               {protocolgroups}
            </Modal>;
 }
+ProtocolViolationModal.propTypes = {
+    URL: PropTypes.string,
+    SeriesUID: PropTypes.string,
+    onClose: PropTypes.func,
+};
 
 /**
  * Return a modal window to display a MRI violations protocol
@@ -289,4 +298,9 @@ function ProtocolCheckViolationModal(props) {
                 Data={data} />
            </Modal>;
 }
+ProtocolCheckViolationModal.propTypes = {
+    URL: PropTypes.string,
+    SeriesUID: PropTypes.string,
+    onClose: PropTypes.func,
+};
 export default ProtocolModal;