+++ tools/single_use/populate_examiners_psc_rel.php
@@ -18,10 +18,10 @@
  * @license  Loris license
  * @link     https://www.github.com/aces/Loris-Trunk/
  */
-require_once 'generic_includes.php';
+require_once __DIR__.'/../generic_includes.php';
 
 $config = NDB_Config::singleton();
-$db     =& Database::singleton();
+$db     = $loris->getDatabaseConnection();
 
 $examinersRows = $db->pselect(
     "SELECT * FROM examiners",