+++ CHANGELOG.md
@@ -13,12 +13,16 @@ changes in the following format: PR #1234***
 #### Features
 - Added new interface intended to be used for querying module data from PHP (PR #8215) 
 - Added the NOT NULL constraint on Project Name (PR #8295)
+- Migrated instrument permissions from config.xml to database and added the ability
+  to manage instrument permissions in the frontend from the `instrument_manager`
+  module. (PR #8302)
 
 #### Updates and Improvements
 - Rename subproject to Cohort (PR #7817)
 - Create new CohortData and CohortController classes to use as data access model 
   and transfer object (PR #7817)
 - BVL Feedback widget only shows notifications for the users sites / projects (PR #7848)
+- Add Date status change value in session table (PR #8350)
 
 #### Bug Fixes
 - placeholder